Move tests alongside code under test
All unit tests have been moved alongside the module under test. This
pattern is generally preferred by the team, and matches how tests are
organized in both the extension and mobile repos.
