+++ src/keyring/KeyringController.test.ts
@@ -6,12 +6,12 @@ import {
   recoverTypedSignatureLegacy,
 } from 'eth-sig-util';
 import { stub } from 'sinon';
-import KeyringController, { Keyring, KeyringConfig } from '../src/keyring/KeyringController';
-import PreferencesController from '../src/user/PreferencesController';
-import ComposableController from '../src/ComposableController';
+import PreferencesController from '../user/PreferencesController';
+import ComposableController from '../ComposableController';
+import KeyringController, { Keyring, KeyringConfig } from './KeyringController';
 
 const Transaction = require('ethereumjs-tx');
-const mockEncryptor: any = require('./utils/mockEncryptor');
+const mockEncryptor: any = require('../../tests/mocks/mockEncryptor');
 
 const input =
   '{"version":3,"id":"534e0199-53f6-41a9-a8fe-d504702ee5e8","address":"b97c80fab7a3793bbe746864db80d236f1345ea7",' +
