+++ src/user/AddressBookController.test.ts
@@ -1,4 +1,4 @@
-import AddressBookController from '../src/user/AddressBookController';
+import AddressBookController from './AddressBookController';
 
 describe('AddressBookController', () => {
   it('should set default state', () => {
