+++ tests/mocks/mockEncryptor.js
