+++ tsconfig.json
@@ -17,6 +17,9 @@
     "target": "es6"
   },
   "include": [
-    "./src/**/*.ts"
+    "src/**/*.ts"
+  ],
+  "exclude": [
+    "**/*.test.ts"
   ]
 }
